Merge commit 'd338abb664febbc2c7266af7818aab1f12dbc161'
* commit 'd338abb664febbc2c7266af7818aab1f12dbc161':
  mpeg4videodec: Call av_log() in check_marker() with AVCodecContext instead of NULL

This commit is a noop.

Since check_marker() is still in get_bits.h in FFmpeg and shared amongst
decoders, a similar but different change will be committed in the next
commit in to serve the same purpose of this one (that is, passing a
logging context to the av_log() in that function).

Merged-by: Clément Bœsch <u@pkh.me>