avformat/Makefile: Remove unnecessary dependencies on vocdec
Commit 0d1229f1d2b8f26dd50c6be7917bb8ed8cb95364 factored the main part
of the voc demuxer's read_packet function out; yet when this Libav
commit was merged in f99195d56f4aab266926724ca1cfae822df4df16, the
dependency of the other users of this function on vocdec.o was
unnecessarily kept. This commit fixes this.

While just at it, also disable the data only used by the voc demuxer
and muxer in voc.c if both of them are disabled.

Reviewed-by: Paul B Mahol <onemda@gmail.com>
Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@gmail.com>