libavformat: add side_data copy in concat demuxer
Adds support for concat demuxer to copy the side data information
from the input file to the resulting file. It will behave like the
metadata copy, where the metadata of the first file is kept in the
the output file.

Extract the current code that already performs the stream side_data
copy into a separate method and reuse the method in the concat demuxer.

Change-Id: I8b8b0da2ce7289a9a83eb3376cb9d04aa71816a7
Signed-off-by: Gerard Sole <g.sole.ca@gmail.com>