FFMPEG: Fix frame pts is invalid when the package contains multiple frames
When a pakage contains multiple frames, only the first frame uses the pts of the pakage, and the pts of the remaining frames are invalid.

In ffplay.c, when the pts of the accepted frame is invalid, the pts of this frame will be calculated based on the nb_samples of this frame and the pts of the previous frame.

Signed-off-by: renjianguang <renjianguang@xiaomi.com>