FFMPEG:  destroy the event resource after the close event returns mediad
Moviesink's close function logic destroyed the event resource,
resulting in the close event not being notified to mediad,
and there was no logic to clear the Opaque user data of the filter.
In the second open command, the correct filter could not be selected because the Opaque user data was not empty.

Signed-off-by: yangsen5 <yangsen5@xiaomi.com>