FFmpeg: set fps analysis framecount to 0 for rtsp
During the find stream_info stage of the RTSP stream, the default
value of `fps_analyze_framecount` is 20, which means that more than
`fps_analyze_framecount` packets need to be read before the correct
`r_frame_rate` and `avg_frame_rate` can be calculated. This process
introduces a significant amount of delay. In reality, the RTSP stream
does not require this condition, as the correct framerate can be
calculated based on the parsed SPS information. Therefore, for RTSP
streams, `fps_analyze_framecount` is set to 0.

Signed-off-by: shizhenghui <shizhenghui@xiaomi.com>