ffmpeg: Add paused to replace underflow logic
During playback, after sending the pause command, when the resume
command is sent again, if the 96K buffer is in the media layer, the
media will continue to fill the buffer until the 96K is full, then it
will enqueue the buffer and send resume when underflow is true. However,
the driver has received the pause command before and stopped data
consumption. It will not notify the media layer when there is underrun,
so underflow is false and the resume command will not be sent to the
driver, resulting in playback pause. To solve this problem, add pause to
replace underflow, and issue the resume command based on the judgment
of pause and dq_count=0.

Signed-off-by: yangsen5 <yangsen5@xiaomi.com>