MAINTAINERS: Add L field based on the linux kernel MAINTAINERs
Text was stolen from the linux kernel
This is thus identical to the kernel just a different more compact format.
I am very happy also to switch the file entirely to the format of the linux kernel maintainer list
if people prefer

This allows specifying a more specific mailing list, if a specific area has
such a list.

Signed-off-by: Michael Niedermayer <michael@niedermayer.cc>