ffmpeg/libavfilter: fix amix could handle quick link.
Here is a circumstance that an input with link 0 sample size could not be freed if state is INPUT_ON | INPUT_EOF | INPUT_BLOCKED， even though it is empty. If its link is set 0 after being set EOF and was pushed some sample into, its state will become to INPUT_EOF and is not empty which will result in failing to request frame and freeze the whole pipeline.

Signed-off-by: wurui3 <wurui3@xiaomi.com>